+++ python/paddle/distributed/fleet/meta_parallel/sharding/group_sharded_stage2.py
@@ -27,6 +27,7 @@ from functools import reduce
 from types import MethodType
 
 import paddle
+import paddle.distributed as dist
 from paddle import nn
 from paddle.distributed import collective
 from paddle.distributed.utils.log_utils import get_logger
@@ -324,12 +325,12 @@ class GroupShardedStage2(nn.Layer):
         """
 
         for buffer in self._layer.buffers(include_sublayers=True):
-            collective.broadcast(
+            dist.broadcast(
                 buffer, self._global_root_rank, self._group, sync_op=True
             )
 
             if self._dp_group and self._dp_group.nranks > 1:
-                collective.broadcast(
+                dist.broadcast(
                     buffer,
                     self._dp_group.ranks[0],
                     self._dp_group,
@@ -402,7 +403,7 @@ class GroupShardedStage2(nn.Layer):
 
                     # Synchronize the reduce parameter gradient asynchronize
                     self._sharding_optimizers[0]._update_task(
-                        collective.reduce(
+                        dist.reduce(
                             tensor=param.grad,
                             dst=self._group.ranks[dst_rank],
                             group=self._group,
@@ -415,7 +416,7 @@ class GroupShardedStage2(nn.Layer):
                             not self._reduce_overlap
                         ), 'dp + stage2 hybrid parallel only Synchronize due to the new communication lib.'
                         # TODO(wuhuachao):after the new communication lib upgrading, overlapping the comm of dp + stage2.
-                        collective.all_reduce(
+                        dist.all_reduce(
                             tensor=param.grad,
                             group=self._dp_group,
                             sync_op=True,
@@ -469,7 +470,7 @@ class GroupShardedStage2(nn.Layer):
                         grad_storage.sent = True
                         # Synchronize the reduce parameter gradient asynchronize
                         self._sharding_optimizers[0]._update_task(
-                            collective.reduce(
+                            dist.reduce(
                                 tensor=grad_storage.buffer,
                                 dst=self._group.ranks[grad_storage.destination],
                                 group=self._group,
@@ -482,7 +483,7 @@ class GroupShardedStage2(nn.Layer):
                                 not self._reduce_overlap
                             ), 'dp + stage2 hybrid parallel only Synchronize due to the new communication lib.'
                             # TODO(wuhuachao):after the new communication lib upgrading, overlapping the comm of dp + stage2.
-                            collective.all_reduce(
+                            dist.all_reduce(
                                 tensor=grad_storage.buffer,
                                 group=self._dp_group,
                                 sync_op=True,
